% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{beams}
\alias{beams}
\title{The beams data set}
\value{
A data frame
}
\description{
Ten wooden beams had
their Strength measured together with their
Gravity and Moisture.
}
\examples{
data(beams, package="jroreillyregression")
}
\keyword{datasets}
