% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chopsticks}
\alias{chopsticks}
\alias{chopsticks_full}
\title{Chopstick efficiency}
\source{
https://bmdatablog.files.wordpress.com/2016/04/chopsticks.pdf
}
\description{
A few researchers set out to determine the optimal length of chopsticks for children and adults.
They came up with a measure of how effective a pair of chopsticks performed, called the "Food Pinching
Performance."
The "Food Pinching Performance" was determined by counting the number of peanuts picked
and placed in a cup (PPPC).

The idea was taken from http://blog.yhat.com/posts/7-funny-datasets.html
}
