% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{correlation}
\alias{correlation}
\title{Spurious Correlations}
\source{
Spurious correlations (http://www.tylervigen.com/view_correlation?id=75)
}
\description{
This data set contains the number of honey producing bee colonies
(Thousands of colonies (USDA)) in the US and the Divorce rate in South Carolina
(Divorces per 1000 people (US Census)). With a correlation of 0.9, this is
clearly a significant result.
}
\examples{
data(correlation)
cor(correlation$bees, correlation$divorce)
}
