% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{facebook}
\alias{facebook}
\title{Facebook metrics Data Set}
\source{
http://archive.ics.uci.edu/ml/datasets/Facebook+metrics
}
\description{
The data is related to posts' published during the year of 2014 on the
Facebook's page of a renowned cosmetics brand. This dataset contains 500 of the 790 rows and
part of the features analyzed by Moro et al. (2016). The remaining were omitted due to confidentiality issues.
Lifetime post total reach	The number of people who saw a page post (unique users).
\describe{
\item{Lifetime post total impressions}{Impressions are the number of times a post from a page is displayed, whether the post is clicked or not. People may see multiple impressions of the same post. For example, someone might see a Page update in News Feed once, and then a second time if a friend shares it.}
\item{Lifetime engaged users}{The number of people who clicked anywhere in a post (unique users).}
\item{Lifetime post consumers}{The number of people who clicked anywhere in a post.}
\item{Lifetime post consumptions}{The number of clicks anywhere in a post.}
\item{Lifetime post impressions by people who have liked a page}{Total number of impressions just from people who have liked a page.}
\item{Lifetime post reach by people who like a page}{The number of people who saw a page post because they have liked that page (unique users).}
\item{Lifetime people who have liked a page and engaged with a post}{The number of people who have liked a Page and clicked anywhere in a post (Unique users).}
\item{Comments}{Number of comments on the publication.}
\item{Likes}{Number of "Likes" on the publication.}
\item{Shares}{Number of times the publication was shared.}
\item{Total interactions}{The sum of "likes", "comments", and "shares" of the post.}
}
}
