% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lsd}
\alias{lsd}
\title{LSD & Maths}
\source{
https://www.ncbi.nlm.nih.gov/pubmed/5676802
}
\description{
Group of volunteers was given LSD, their mean scores on math exam and
tissue concentrations of LSD were obtained at n=7 time points.

The test score is out of 100 and the Drugs is (ppm).
}
