% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{rugby}
\alias{rugby}
\title{Rugby players sizes}
\source{
http://www.bbc.co.uk/blogs/tomfordyce/2012/03/land_of_the_rugby_giants.html
}
\description{
Sizes of the England XV to the equivalent week in the Five/Six Nations of in 1962, 1972, ..., 2012.
}
