% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{starbucks}
\alias{starbucks}
\title{The nutritional value of Starbucks}
\source{
https://www.kaggle.com/starbucks/
}
\description{
This dataset includes the nutritional information for Starbucks'
food menu items.
}
